nasdaq_scanner/requirements.txt
# NASDAQ Volatility Scanner - Dependencies

# Market data
alpaca-trade-api>=3.0.0
yfinance>=0.2.36

# Data analysis
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0

# Technical analysis
ta>=0.10.2

# HTTP and webhooks
requests>=2.28.0

# Environment management
python-dotenv>=1.0.0

# Scheduling (optional)
schedule>=1.2.0

# Logging
colorlog>=6.7.0
